+++ server/src/java/org/jppf/server/nio/nodeserver/TaskQueueChecker.java
@@ -152,7 +152,7 @@ public class TaskQueueChecker implements Runnable
 			}
 			acceptableChannels.add(i);
 		}
-		for (Integer i: channelsToRemove) idleChannels.remove(i);
+		for (Integer i: channelsToRemove) idleChannels.remove(i.intValue());
 		if (debugEnabled) log.debug("found " + acceptableChannels.size() + " acceptable channels");
 		if (!acceptableChannels.isEmpty())
 		{
